server/requirements.txt
Django>=4.2
djangorestframework
django-cors-headers

django-storages
boto3
